requirements.txt
altgraph==0.17.3
certifi==2022.9.24
charset-normalizer==2.1.1
future==0.18.2
idna==3.4
imutils==0.5.4
numpy==1.23.4
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
pefile==2022.5.30
Pillow==9.3.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pywin32-ctypes==0.2.0
requests==2.28.1
scipy==1.9.3
torch==1.13.0
torch-mtcnn==0.0.7
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.12

configparser~=5.3.0
pyttsx3~=2.90